Review 5/6 — move the CLI modules that outlive `storybook ai` into the tools CLI
The tools CLI reached into `cli/ai/mcp/` for instance discovery, config-dir
resolution and the schema help renderer, so the surviving code lived inside
the tree that is slated for deletion. Relocate everything both CLIs need into
`cli/tools/` and have `storybook ai` import it from there, so removing the ai
CLI later is a pure deletion of its directory plus its registration lines.

  ai/mcp/registry.ts         -> tools/instances/registry.ts
  ai/mcp/resolve-instance.ts -> tools/instances/resolve.ts
  ai/mcp/types.ts            -> split: instance records to tools/instances/types.ts,
                                MCP wire shapes into tools/mcp-client.ts,
                                JSON Schema nodes to tools/schema-lines.ts
  ai/mcp/client.ts           -> tools/mcp-client.ts
  resolveStorybookConfigDir  -> tools/config-dir.ts
  schemaLines                -> tools/schema-lines.ts
  CommandFailureHandler      -> tools/register.ts

Tests move with their modules. Behavior-neutral: no runtime change on either
CLI, and no `cli/tools` -> `cli/ai` import remains.
