+++ code/addons/mcp/src/constants.ts
@@ -2,10 +2,10 @@ export const MCP_APP_PARAM = 'mcp-app';
 export const MCP_APP_SIZE_CHANGED_EVENT = 'storybook-mcp:size-changed';
 
 /**
- * Request header the `storybook ai` CLI (and the Claude/Codex plugins built on
- * it) sends on every MCP request to mark itself as a trusted local Storybook
- * client. Mirrors `STORYBOOK_MCP_PROXY_HEADER` in Storybook core's
- * `cli/ai/mcp/client.ts` — the two must stay in sync.
+ * Request header Storybook's own CLIs (and the Claude/Codex plugins built on
+ * them) send on every MCP request to mark themselves as a trusted local
+ * Storybook client. Mirrors `STORYBOOK_MCP_PROXY_HEADER` in Storybook core's
+ * `cli/tools/mcp-client.ts` — the two must stay in sync.
  */
 export const STORYBOOK_MCP_PROXY_HEADER = 'X-Storybook-MCP-Proxy';
