+++ code/core/src/cli/ai/mcp/local-metadata.test.ts
@@ -4,7 +4,7 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { experimental_loadStorybook as loadStorybook } from 'storybook/internal/core-server';
 
-import { loadStorybookAiMetadata, resolveStorybookConfigDir } from './local-metadata.ts';
+import { loadStorybookAiMetadata } from './local-metadata.ts';
 
 vi.mock('storybook/internal/core-server', { spy: true });
 
@@ -123,23 +123,5 @@ describe('loadStorybookAiMetadata', () => {
     mockLoadedStorybook(vi.fn().mockResolvedValue(undefined));
 
     await expect(loadStorybookAiMetadata({ cwd: '/repo' })).resolves.toBeUndefined();
-  });
-});
-
-describe('resolveStorybookConfigDir', () => {
-  it('defaults to .storybook under the target cwd', () => {
-    expect(resolveStorybookConfigDir({ cwd: '/repo' })).toBe(resolve('/repo/.storybook'));
-  });
-
-  it('resolves relative config dirs from the target cwd', () => {
-    expect(resolveStorybookConfigDir({ cwd: '/repo', configDir: 'config/storybook' })).toBe(
-      resolve('/repo/config/storybook')
-    );
-  });
-
-  it('keeps absolute config dirs unchanged', () => {
-    expect(resolveStorybookConfigDir({ cwd: '/repo', configDir: '/custom/.storybook' })).toBe(
-      '/custom/.storybook'
-    );
   });
 });
