+++ code/core/src/cli/ai/mcp/run-tool.ts
@@ -3,24 +3,27 @@ import { resolve } from 'node:path';
 import * as v from 'valibot';
 
 import { detectAgent } from '../../../telemetry/detect-agent.ts';
-import { McpJsonRpcError, callMcpTool, listMcpTools } from './client.ts';
-import { getInterceptMarkdown } from './intercepts.ts';
+import { resolveStorybookConfigDir } from '../../tools/config-dir.ts';
+import { readRegistry } from '../../tools/instances/registry.ts';
+import { resolveInstance } from '../../tools/instances/resolve.ts';
+import type { InterceptReason, StorybookInstanceRecord } from '../../tools/instances/types.ts';
 import {
-  loadStorybookAiMetadata,
-  resolveStorybookConfigDir,
-  type StorybookAiMetadata,
-} from './local-metadata.ts';
-import { readRegistry } from './registry.ts';
-import { resolveInstance } from './resolve-instance.ts';
+  McpJsonRpcError,
+  ToolCallResultSchema,
+  callMcpTool,
+  listMcpTools,
+  type McpToolDescriptor,
+  type ToolCallResult,
+} from '../../tools/mcp-client.ts';
+import {
+  JsonSchemaNodeSchema,
+  MAX_SCHEMA_DEPTH,
+  schemaLines,
+  type JsonSchemaNode,
+} from '../../tools/schema-lines.ts';
+import { getInterceptMarkdown } from './intercepts.ts';
+import { loadStorybookAiMetadata, type StorybookAiMetadata } from './local-metadata.ts';
 import { parsePort, parseToolArgs } from './tool-args.ts';
-import type {
-  InterceptReason,
-  JsonSchemaNode,
-  McpToolDescriptor,
-  StorybookInstanceRecord,
-  ToolCallResult,
-} from './types.ts';
-import { JsonSchemaNodeSchema, ToolCallResultSchema } from './types.ts';
 
 /**
  * Why an invocation failed before any command executed, for the `ai-command` telemetry event
@@ -542,70 +545,6 @@ function formatToolResult(result: ToolCallResult): string {
 
 function getLocalToolNames(metadata: StorybookAiMetadata): Set<string> {
   return new Set(Object.keys(metadata.localTools ?? {}));
-}
-
-export const MAX_SCHEMA_DEPTH = 4;
-
-function schemaItem(schema: JsonSchemaNode): JsonSchemaNode | undefined {
-  return Array.isArray(schema.items) ? schema.items[0] : schema.items;
-}
-
-function schemaTypeLabel(schema: JsonSchemaNode): string | undefined {
-  if (schema.type === 'array') {
-    const item = schemaItem(schema);
-    return item?.type ? `array of ${item.type}` : 'array';
-  }
-  if (schema.anyOf || schema.oneOf) {
-    return 'one of';
-  }
-  return schema.type;
-}
-
-function schemaChildLines(schema: JsonSchemaNode, indent: string, depth: number): string[] {
-  if (depth <= 0) {
-    return [];
-  }
-  const lines: string[] = [];
-
-  if (schema.type === 'object' && schema.properties) {
-    const required = new Set(schema.required ?? []);
-    for (const [name, child] of Object.entries(schema.properties)) {
-      lines.push(...schemaLines(`\`${name}\``, child, required.has(name), indent, depth));
-    }
-  }
-
-  const item = schemaItem(schema);
-  if (item && ((item.type === 'object' && item.properties) || item.anyOf || item.oneOf)) {
-    lines.push(`${indent}each item:`);
-    lines.push(...schemaChildLines(item, `${indent}  `, depth - 1));
-  }
-
-  const variants = schema.anyOf ?? schema.oneOf;
-  if (variants) {
-    variants.forEach((variant, index) => {
-      const suffix = variant.description ? `: ${variant.description}` : '';
-      lines.push(`${indent}option ${index + 1}${suffix}`);
-      lines.push(...schemaChildLines(variant, `${indent}  `, depth - 1));
-    });
-  }
-
-  return lines;
-}
-
-/** Also the `storybook tools` help renderer's line formatter, so both CLIs render schemas alike. */
-export function schemaLines(
-  label: string,
-  schema: JsonSchemaNode,
-  isRequired: boolean,
-  indent: string,
-  depth: number
-): string[] {
-  const meta = [schemaTypeLabel(schema), isRequired ? 'required' : undefined]
-    .filter(Boolean)
-    .join(', ');
-  const description = schema.description ? `: ${schema.description}` : '';
-  const head = `${indent}- ${label}${meta ? ` (${meta})` : ''}${description}`;
-  return [head, ...schemaChildLines(schema, `${indent}  `, depth - 1)];
 }
 
 function formatToolHelp(tool: McpToolDescriptor, { local }: { local: boolean }): string {
