+++ code/core/src/cli/tools/bootstrap.ts
@@ -22,7 +22,7 @@ import type {
   ToolsetGetService,
 } from '../../shared/open-service/toolset-definition.ts';
 import { getRegisteredToolsets } from '../../shared/open-service/toolset-registry.ts';
-import { resolveStorybookConfigDir } from '../ai/mcp/local-metadata.ts';
+import { resolveStorybookConfigDir } from './config-dir.ts';
 import type { ToolsTarget } from './discover-instance.ts';
 
 export type ToolsRuntime = {
