+++ code/core/src/cli/tools/instances/registry.test.ts
