+++ code/core/src/cli/tools/instances/registry.ts
