+++ code/core/src/cli/tools/instances/resolve.ts
